requirements.txt
altgraph==0.17
autopep8==1.5.7
future==0.18.2
numpy==1.20.3
pefile==2021.5.24
pillow==8.2.0
pycodestyle==2.7.0
pygame==2.0.1
pyinstaller==4.3
pyinstaller-hooks-contrib==2021.1
pywin32-ctypes==0.2.0
toml==0.10.2
